Разобраться бл*ть с группами

+++ Plugin_for_Pioneer/Import/Import.cs
@@ -93,7 +93,7 @@ namespace Plugin_for_Pioneer
 
                 foreach (var seleсtedElement in selectedRef)
                 {
-                    Element element = doc.GetElement(seleсtedElement);                 
+                    Element element = doc.GetElement(seleсtedElement);
 
                     if ((BuiltInCategory)element.Category.Id.IntegerValue == BuiltInCategory.OST_IOSModelGroups ||
                         (BuiltInCategory)element.Category.Id.IntegerValue == BuiltInCategory.OST_Sections ||
@@ -132,32 +132,11 @@ namespace Plugin_for_Pioneer
                 {
                     categorySet.Insert(Category.GetCategory(doc, category));
                 }
-               
+
                 //IEnumerable<Element> elementListDistinct = elementList.Distinct();
                 using (Transaction ts = new Transaction(doc, "Add parameter"))
                 {
                     ts.Start();
-                    /*foreach (var element in elementListDistinct)
-                    {
-                        if (element.LookupParameter("PNR_Код по классификатору") == null
-                            || element.LookupParameter("PNR_Описание по классификатору") == null)
-                        {
-                            CreateShared createShared_pnr_1 = new CreateShared();
-                            createShared_pnr_1.CreateSharedParameter(uiapp.Application,
-                                                               doc,
-                                                               "PNR_Код по классификатору",
-                                                               categorySet,
-                                                               BuiltInParameterGroup.PG_IDENTITY_DATA,
-                                                               true);
-                            CreateShared createShared_pnr_2 = new CreateShared();
-                            createShared_pnr_2.CreateSharedParameter(uiapp.Application,
-                                                               doc,
-                                                               "PNR_Описание по классификатору",
-                                                               categorySet,
-                                                               BuiltInParameterGroup.PG_IDENTITY_DATA,
-                                                               true);
-                        }
-                    }*/
                     CreateShared createShared_pnr_1 = new CreateShared();
                     createShared_pnr_1.CreateSharedParameter(uiapp.Application,
                                                        doc,
@@ -186,9 +165,9 @@ namespace Plugin_for_Pioneer
                     var element = listDataElementSorted.FirstOrDefault(x => x.guid == listDataExcelSorted[X].guid);
 
                     if (element != null)
-                    { 
+                    {
                         //Если парамтеры не равны, добавляем в массив изменяемых элементов
-                        if (element.pnr_1 != listDataExcelSorted[X].pnr_1 
+                        if (element.pnr_1 != listDataExcelSorted[X].pnr_1
                         || element.pnr_2 != listDataExcelSorted[X].pnr_2)
                         {
                             if (listDataExcelSorted[X].pnr_1 != "" & listDataExcelSorted[X].pnr_2 != "")
@@ -197,13 +176,16 @@ namespace Plugin_for_Pioneer
                     }
                 });
 
-                //Заносим значения в параметр у элементов с флагом
-
                 List<List<ElementId>> groupElements = new List<List<ElementId>>();
                 List<String> groupNames = new List<String>();
                 List<Element> groupId = new List<Element>();
+                List<GroupType> groupTypes = new List<GroupType>();
+
                 var listDataElementTrueSorted = listDataElementTrue.OrderBy(pr => pr.guid).ToList();
 
+                List<Group> groups = new List<Group>();
+
+                //Заносим значения в параметр из листа нужных элементов 
                 if (listDataElementTrue.Count > 0)
                 {
                     Transaction transaction = new Transaction(doc, "Заносим значения в параметр");
@@ -213,40 +195,52 @@ namespace Plugin_for_Pioneer
                         if (desieredElementTrue == null)
                             continue;
                         var excelElement = listDataExcel.FirstOrDefault(r => r.guid == desieredElementTrue.guid);
-                        if (desieredElementTrue != null) 
+                        if (desieredElementTrue != null)
                         {
-                            if(desieredElementTrue.element.GroupId.IntegerValue != -1)
+                            if (desieredElementTrue.element.GroupId.IntegerValue != -1)
                             {
                                 Group group = (Group)doc.GetElement(desieredElementTrue.element.GroupId);
-                                
+
                                 if (group.GroupId.IntegerValue != -1)
                                     continue;
-                                
-                                groupId.Add(doc.GetElement(desieredElementTrue.element.GroupId));
-                                groupNames.Add(group.Name);
-                                group.GroupType.Name = "DELETE";
-                                List<ElementId> elements = group.UngroupMembers().ToList();
-                                doc.Delete(group.Id);
-                                groupElements.Add(elements);
+
+                                /*                                groupId.Add(doc.GetElement(desieredElementTrue.element.GroupId));
+                                                                groupNames.Add(group.Name);
+                                                                groupTypes.Add(group.GroupType);
+                                                                group.GroupType.Name = "DELETE";
+                                                                List<ElementId> elements = group.UngroupMembers().ToList();
+                                                                doc.Delete(group.Id);
+                                                                groupElements.Add(elements);*/
+
+
+                                if(groups.Contains(group) == false)
+                                {
+                                    groups.Add(group);
+                                }
                             }
+                            //Заносим значение в параметр
                             desieredElementTrue.element.LookupParameter("PNR_Код по классификатору").Set(excelElement.pnr_1);
                             desieredElementTrue.element.LookupParameter("PNR_Описание по классификатору").Set(excelElement.pnr_2);
                         }
                     }
+                    //PurgeDocument.Purge(doc);
                     transaction.Commit();
                 }
 
-                Transaction tr = new Transaction(doc, "NewGroup");
-                tr.Start();
-                int i = 0;
-                foreach (var group in groupElements)
-                {
-                    Group groupNew = doc.Create.NewGroup(group);
-                    groupNew.GroupType.Name = groupNames[i];
-                    i++;
-                }
-                PurgeDocument.Purge(doc);
-                tr.Commit();
+                /*                Transaction tr = new Transaction(doc, "NewGroup");
+                                tr.Start();
+                                int i = 0;
+                                foreach (var groupNewInOld in groupElements)
+                                {
+                                    Group groupNew = doc.Create.NewGroup(groupNewInOld);
+                                    groupNew.GroupType = groupTypes[i];
+                                    //groupNew.GroupType.Name = groupNames[i];
+                                    i++;
+                                }
+                                //PurgeDocument.Purge(doc);
+                                tr.Commit();*/
+
+                TaskDialog.Show("x", $"{groups.Count()}");
             }
 
             catch (Autodesk.Revit.Exceptions.OperationCanceledException) { }
@@ -399,12 +393,15 @@ namespace Plugin_for_Pioneer
             //Attempting to recover all purgeable elements and delete them from the document
             List<ElementId> purgeableElementIds = GetPurgeableElements(doc, performanceAdviserRuleIds);
 
-            foreach (var purgeableElementId in purgeableElementIds)
+            if (purgeableElementIds.Count > 0)
             {
-                if (purgeableElementId != null)
+                foreach (var purgeableElementId in purgeableElementIds)
                 {
-                    if (doc.GetElement(purgeableElementId).Name == "DELETE")
-                        doc.Delete(purgeableElementId);
+                    if (purgeableElementId != null)
+                    {
+                        if (doc.GetElement(purgeableElementId).Name == "DELETE")
+                            doc.Delete(purgeableElementId);
+                    }
                 }
             }
         }
